+++ src/mip.h
@@ -10,4 +10,4 @@ struct mip_header {
 
 struct mip_header *create_arp_request_package(uint8_t src_addr);
 
-struct mip_header *create_arp_response_package(int src_addr, struct mip_header *arp_request);
+struct mip_header *create_arp_response_package(uint8_t src_addr, struct mip_header *arp_request);
